Add files via upload
This is a quick and dirty "conditional formatting" in sheets. I had two sheets and needed to highlight the cells in a column that contained a string found in another column. 
There might be a more elegant solution builtin to Sheets but I had hard time finding one so I wrote this.